requirements.txt
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.18
certifi==14.5.14
Django==1.6.1
django-celery==3.1.16
django-countries==3.0.1
django-crontab==0.6.0
django-encrypted-fields==1.1.1
django-file-resubmit==0.4.1
django-simple-captcha==0.4.3
django-social-auth==0.7.28
django-socketio==0.3.8
flup==1.0.2
gevent==0.13.6
gevent-socketio==0.2.1
gevent-websocket==0.2.3
greenlet==0.4.5
html5lib==0.999
httplib2==0.9
kombu==3.0.26
oauth2==1.5.211
PIL==1.1.7
Pillow==2.6.1
pisa==3.0.33
psycopg2==2.5.4
pyasn1==0.1.7
pycrypto==2.6.1
python-dateutil==2.2
python-keyczar==0.71rc0
python-openid==2.2.5
pytz==2015.4
redis==2.10.3
reportlab==3.1.44
six==1.8.0
sockjs-tornado==1.0.1
South==1.0.1
sphinx-me==0.3
SwampDragon==0.3.8
tornado==4.0.2
tornado-redis==2.4.18
